Implement KeyboardInput event source
The KeyboardInput event source allows users to retrieve input from STDIN
as soon as it arrives. By default, the event source creates an
std::string containing the input data without the trailing newline, if
there was any.

This event source is now used to allow users of the Gateway server to
type "exit" in the terminal and terminate the server gracefully.